Update the merging of sound and sensor code
The sound playing code now "works" in that it can play a sound when the distance
detector detects movement into the detection zone. However, it quickly gets
overwhelmed as it tries to play multiple times to the same stream while a
previous sound is still playing there.

TODO: Find a way to only play a sound on a stream that is not currently playing.
(Which will also preferably involve 'debouncing' the sound playing function).