Convert current demo samples to 48k
This is because the USB audio devices currently only allow playback at 48k.
The sounddevice doesn't provide an easy way to resample the sounds on the fly
either - so the simplest solution for the moment is to convert the sample rate
of the samples right from the beginning.

If we ever find more flexible audio devices that can play back at multiple sample
rates, it should be simple enough to re-introduce samples that have multiple
sample rates.